Add dependent packages to gmr DESCRIPTION file.
Added gdata and shiny to Imports in DESCRIPTION file.
The packages will now be automatically downloaded when gmr is
sourced from github or built locally. I am unsure if there are other
packages needed. If a few functions are needed then they should be
included in Imports. If the entire package should be available
then they can be included in Depends. Though I think better code
is to explicitly state in the gmr function which packages the function
depends upon and then keep that list in Imports.